feat(sidebar) Add tests for sidebar widget intents
